requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
loguru
pyserial
